changes to lesson 2 snippets
