+++ tutorials/nservicebus-step-by-step/1-getting-started/Snippets/Core_9/Lesson1.cs
@@ -53,6 +53,9 @@ class StepByStep
         #region Startup
         builder.UseNServiceBus(endpointConfiguration);
 
+        Console.WriteLine("Press Enter to exit...");
+        Console.ReadLine();
+
         await builder.Build().RunAsync();
         #endregion
     }
