--- tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Core_8/Command.cs
@@ -1,14 +1,0 @@
-﻿using NServiceBus;
-
-namespace Core_8
-{
-    #region Command
-
-    public class DoSomething :
-        ICommand
-    {
-        public string SomeProperty { get; set; }
-    }
-
-    #endregion
-}
