--- tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Core_8/ComplexCommand.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections.Generic;
-using NServiceBus;
-
-namespace Core_8
-{
-    #region ComplexCommand
-
-    public class DoSomethingComplex :
-        ICommand
-    {
-        public int SomeId { get; set; }
-        public ChildClass ChildStuff { get; set; }
-        public List<ChildClass> ListOfStuff { get; set; } = new List<ChildClass>();
-    }
-
-    public class ChildClass
-    {
-        public string SomeProperty { get; set; }
-    }
-
-    #endregion
-}
