--- tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Core_8/EmptyHandler.cs
@@ -1,17 +1,0 @@
-﻿using System.Threading.Tasks;
-using NServiceBus;
-
-namespace Core_8.EmptyHandler
-{
-    #region EmptyHandler
-    public class DoSomethingHandler :
-        IHandleMessages<DoSomething>
-    {
-        public Task Handle(DoSomething message, IMessageHandlerContext context)
-        {
-            // Do something with the message here
-            return Task.CompletedTask;
-        }
-    }
-    #endregion
-}
