+++ tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Core_9/Command.cs
@@ -1,0 +1,13 @@
+﻿using NServiceBus;
+
+namespace Core_9;
+
+#region Command
+
+public class DoSomething :
+    ICommand
+{
+    public string SomeProperty { get; set; }
+}
+
+#endregion
