+++ tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Core_9/ComplexCommand.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections.Generic;
+using NServiceBus;
+
+namespace Core_9;
+
+#region ComplexCommand
+
+public class DoSomethingComplex :
+    ICommand
+{
+    public int SomeId { get; set; }
+    public ChildClass ChildStuff { get; set; }
+    public List<ChildClass> ListOfStuff { get; set; } = new List<ChildClass>();
+}
+
+public class ChildClass
+{
+    public string SomeProperty { get; set; }
+}
+
+#endregion
