+++ tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Core_9/EmptyHandler.cs
@@ -1,0 +1,16 @@
+﻿using System.Threading.Tasks;
+using NServiceBus;
+
+namespace Core_9.EmptyHandler;
+
+#region EmptyHandler
+public class DoSomethingHandler :
+    IHandleMessages<DoSomething>
+{
+    public Task Handle(DoSomething message, IMessageHandlerContext context)
+    {
+        // Do something with the message here
+        return Task.CompletedTask;
+    }
+}
+#endregion
