+++ tutorials/nservicebus-step-by-step/2-sending-a-command/Snippets/Snippets.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29728.190
 MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core_8", "Core_8\Core_8.csproj", "{10CE989A-8E4F-400C-8BBD-3935B9FEE695}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core_9", "Core_9\Core_9.csproj", "{10CE989A-8E4F-400C-8BBD-3935B9FEE695}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
